% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atualizar_dados.R
\name{atualizar_dados}
\alias{atualizar_dados}
\title{Funcao para atualizar dados de mananciais}
\usage{
atualizar_dados()
}
\value{
Uma tibble
}
\description{
Usando os dados disponíveis no dados_mananciais(),
confere quais são os dias faltantes e busca os mesmos
}
\examples{
atualizar_dados()
}
