% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atualizar_dados.R
\name{checar_se_necessario_atualizar}
\alias{checar_se_necessario_atualizar}
\title{Funcao para checar se é necessário atualizar dados de mananciais}
\usage{
checar_se_necessario_atualizar()
}
\value{
Uma tibble
}
\description{
Usando os dados disponíveis no dados_mananciais(),
confere é necessário atualizar a base
}
\examples{
checar_se_necessario_atualizar()
}
\keyword{internal}
