% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpar_dados.R
\name{limpar_dados}
\alias{limpar_dados}
\title{Title}
\usage{
limpar_dados(
  ano,
  diretorio_download = "data-raw/",
  diretorio_salvar = "data-raw/data-clean/"
)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Title
}
