% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_ano.R
\name{obter_ano}
\alias{obter_ano}
\title{Obtém dados de um ano.}
\usage{
obter_ano(
  ano,
  ano_atual = FALSE,
  download_local = TRUE,
  diretorio_download = "data-raw/"
)
}
\arguments{
\item{ano}{Numérico. Qual é o ano que deve ser feito o download. Ex: 2019.}

\item{ano_atual}{Lógico. Se quiser obter dados de anos anteriores, deve ser FALSE. Se quiser obter dados do ano atual, deve ser TRUE. Por padrão, é FALSE.}

\item{download_local}{Lógico. Se for TRUE, uma cópia dos dados baixados será salvo localmente.}

\item{diretorio_download}{Diretório onde o arquivo final deve ser salvo. Ex: \code{"data-raw/"}}
}
\value{
Retorna um tibble com os dados do ano.
}
\description{
A tibble que retorna é a versão original obtida na API, ou seja, não foi limpa.
#' Essa função demora alguns minutos para executar.
}
\examples{
 # mananciais:::obter_ano(2019, download_local = FALSE)
 # mananciais:::obter_ano(2020, ano_atual = TRUE, download_local = FALSE)
}
