% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raspar_site.R
\name{raspar_site}
\alias{raspar_site}
\title{Title}
\usage{
raspar_site(
  ano,
  ano_atual = FALSE,
  diretorio_download = "data-raw/",
  download_local = TRUE
)
}
\arguments{
\item{ano_atual}{}
}
\value{

}
\description{
Title
}
