% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unir_dados.R
\name{unir_dados}
\alias{unir_dados}
\title{Title}
\usage{
unir_dados(
  anos,
  diretorio_dados = "data-raw/data-clean/",
  diretorio_salvar = "data/"
)
}
\arguments{
\item{diretorio_salvar}{}
}
\value{

}
\description{
Title
}
