% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_dia.R
\name{obter_dia}
\alias{obter_dia}
\title{Obtém dados de um único dia.}
\usage{
obter_dia(dia)
}
\arguments{
\item{dia}{Uma string com data, no formato ano-mês-dia. Ex: "2020-09-27"}
}
\value{
Retorna uma tibble com dados dos mananciais para o dia informado
}
\description{
A tibble que retorna é a versão original obtida na API, ou seja, não foi limpa.
}
\examples{
mananciais:::obter_dia("2020-09-26")

}
