% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unir_anos.R
\name{unir_anos}
\alias{unir_anos}
\title{Unir bases de anos}
\usage{
unir_anos(
  anos,
  usar_baixados = TRUE,
  diretorio_dados = "data-raw/data-clean/",
  salvar_local = FALSE,
  diretorio_salvar = "data/"
)
}
\arguments{
\item{anos}{Quais anos deseja unir. Ex: \code{2000:2019}}

\item{usar_baixados}{Lógico. Caso \code{usar_baixados = TRUE}, utilizará os dados baixados anteriormente.}

\item{salvar_local}{Lógico. Caso \code{salvar_local = TRUE}, irá salvar o resultado no diretório informado no argumento \code{diretorio_salvar}}

\item{diretorio_salvar}{Texto. OPCIONAL. Local onde salvar os arquivos limpos, caso \code{salvar_local = TRUE}. Ex: \code{"data/"}}

\item{diretorio_dados.}{Caso \code{usar_baixados = TRUE}, utilizará os dados baixados anteriormente. Informar em qual diretório os arquivos estão salvos.  Ex: \code{"data-raw/data-clean/"}}
}
\value{
Retorna um tibble com os dados unidos.
}
\description{
Usa os dados salvos localmente. A função não está adaptada para usar com pipe ( \%>\% ).
}
\examples{
# mananciais:::unir_anos(2000:2019)

}
